Refactor pinot-thrift
