+++ pom.xml
@@ -58,6 +58,7 @@
     <module>pinot-record-readers/pinot-json</module>
     <module>pinot-record-readers/pinot-orc</module>
     <module>pinot-record-readers/pinot-parquet</module>
+    <module>pinot-record-readers/pinot-thrift</module>
     <module>pinot-connectors</module>
   </modules>
